+++ packages/web-frontend/src/components/View/utils.js
@@ -5,7 +5,7 @@
 
 import React from 'react';
 import { VIEW_STYLES } from '../../styles';
-import { Chart } from './Chart';
+import { ChartWrapper } from './ChartWrapper';
 import { ColorListWrapper } from './ColorListWrapper';
 import { DataDownloadWrapper } from './DataDownloadWrapper';
 import { ListWrapper } from './ListWrapper';
@@ -43,7 +43,7 @@ const VIEW_TYPES = {
 export const getViewWrapper = ({ type, viewType }) => {
   switch (type) {
     case 'chart':
-      return Chart;
+      return ChartWrapper;
     case 'matrix':
       return MatrixWrapper;
     default:
